TVB-2489. Implement logic to deserialize model state_variables_range which are declared as Final attrs.
